[keytap2-gui] Some major changes:
- Adding new approach - Non-exact substitution cipher attack:
Simulated annealing optimization of a combined clustering and language model
cost function. Seems to produce better results in most cases.
Currently it is quite slow. Need to reduce the number of free parameters.
The cost function is not normalized - have to figure out a way to do that.

- Add automatic 'Space' detection:
Use a modified subbreak cost function to try detect word breaks. Need to
add an option to listen quickly to the detected spaces in order to analyze if
the automatic detection was good enough.